Correct a serious bug in squishyball stereo downmix that was blackholing entire channels.
git-svn-id: https://svn.xiph.org/trunk/squishyball@19022 0101bb08-14d6-0310-b084-bc0e0c8e3800

+++ audio.c
@@ -163,7 +163,7 @@ float convert_to_mono(pcm_t *pcm){
 }
 
 /* non-normalized */
-static const int left_mix[33]={
+static const float left_mix[33]={
   1.,    /* A: M */
   1.,    /* B: L */
   0.,    /* C: R */
@@ -179,7 +179,7 @@ static const int left_mix[33]={
   0
 };
 
-static const int right_mix[33]={
+static const float right_mix[33]={
   1.,    /* A: M */
   0.,    /* B: L */
   1.,    /* C: R */
